change names of grid hero
